Add connect code lens
